+++ src/templates/np.plugin.starter/src/NPTriggers-Hooks.js
@@ -4,7 +4,7 @@
 import pluginJson from '../plugin.json' // gives you access to the contents of plugin.json
 import { log, logError, logDebug, timer, clo, JSP } from '@helpers/dev'
 import { updateSettingData, pluginUpdated } from '@helpers/NPConfiguration'
-import { editSettings } from '@helpers/NPSettings'
+import { showMessage } from '@helpers/userInput'
 
 /**
  * NOTEPLAN PER-NOTE TRIGGERS
@@ -107,14 +107,11 @@ export async function onSettingsUpdated(): Promise<void> {
 }
 
 /**
- * Update Settings/Preferences (for iOS etc)
- * Plugin entrypoint for command: "/{{pluginName}}: Update Plugin Settings"
- * @author @dwertheimer
+ * Check the version of the plugin (and force an update if the version is out of date)
  */
-export async function updateSettings() {
+export async function versionCheck(): Promise<void> {
   try {
-    logDebug(pluginJson, `updateSettings running`)
-    await editSettings(pluginJson)
+    await showMessage(`Current Version: ${pluginJson['plugin.version']}`, 'OK', `${pluginJson['plugin.name']}`, true)
   } catch (error) {
     logError(pluginJson, JSP(error))
   }